build: better metadata
