+++ packages/browserless/package.json
@@ -11,7 +11,7 @@
   },
   "repository": {
     "type": "git",
-    "url": "git+https://github.com/kikobeats/browserless.git"
+    "url": "https://github.com/kikobeats/browserless/tree/master/packages/browserless"
   },
   "bugs": {
     "url": "https://github.com/Kikobeats/browserless/issues"
@@ -21,6 +21,7 @@
     "browserless",
     "chrome",
     "chromeless",
+    "core",
     "headless",
     "html",
     "pdf",
@@ -36,17 +37,14 @@
     "require-one-of": "~1.0.2"
   },
   "devDependencies": {
+    "@browserless/test": "latest",
     "ava": "latest",
     "coveralls": "latest",
-    "img-diff-js": "latest",
-    "is-ci": "latest",
     "nyc": "latest",
-    "pdf-parse": "latest",
-    "puppeteer": "latest",
-    "temperment": "latest"
+    "puppeteer": "latest"
   },
   "engines": {
-    "node": ">= 6"
+    "node": ">= 8"
   },
   "files": [
     "src"
