+++ packages/examples/src/screenshot.js
