+++ packages/examples/src/text.js
