+++ packages/pool/test/index.js
@@ -1,7 +1,1 @@
-'use strict'
-
-const path = require('path')
-
-const spec = path.resolve(__dirname, '../../browserless/test/spec.js')
-
-require(spec)(require('../src'))
+require('@browserless/test')(require('../src'))
