+++ packages/test/fixtures/example.jpeg
