+++ packages/test/index.js
@@ -56,20 +56,4 @@ module.exports = createBrowserless => {
     const data = await pdf(buffer)
     t.snapshot(data.text.trim())
   })
-  ;['pdf', 'screenshot', 'html', 'text'].forEach(method => {
-    test(`.${method} wrap errors`, async t => {
-      const browserless = createBrowserless()
-      const timeout = 500
-
-      const error = await t.throwsAsync(browserless[method]('https://example.com', { timeout }))
-
-      t.is(error.name, 'TimeoutError')
-      t.is(error.message, `Navigation Timeout Exceeded: ${timeout}ms exceeded`)
-
-      const browser = await browserless.browser
-      const pages = await browser.pages()
-
-      t.is(pages.length, 1) // about:page is always open
-    })
-  })
 }
